requirements.txt
Flask
python-dotenv
openai
requests
markdown
WeasyPrint